+++ isort/main.py
@@ -505,13 +505,6 @@ def parse_args(argv: Optional[Sequence[str]] = None) -> Dict[str, Any]:
         help="Tells isort to apply changes recursively without asking",
     )
     parser.add_argument(
-        "--unsafe",
-        dest="unsafe",
-        action="store_true",
-        help="Tells isort to look for files in standard library directories, etc. "
-        "where it may not be safe to operate in",
-    )
-    parser.add_argument(
         "--case-sensitive",
         dest="case_sensitive",
         action="store_true",
@@ -544,8 +537,6 @@ def parse_args(argv: Optional[Sequence[str]] = None) -> Dict[str, Any]:
     arguments = {key: value for key, value in vars(parser.parse_args(argv)).items() if value}
     if "dont_order_by_type" in arguments:
         arguments["order_by_type"] = False
-    if arguments.pop("unsafe", False):
-        arguments["safety_excludes"] = False
     return arguments