+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Changelog
     - Since there is no longer composition negative form settings (such as --dont-skip) are no longer required and have been removed.
   - Two-letter shortened setting names (like `ac` for `atomic`) now require two dashes to avoid ambiguity: `--ac`.
   - `length_sort_{section_name}` config usage has been deprecated. Instead `length_sort_sections` list can be used to specify a list of sections that need to be length sorted.
+  - `safety_excludes` and `unsafe` have been deprecated
+  - Config now includes as default full set of safety directories defined by safety excludes.
 
 Internal:
   - isort now utilizes mypy and typing to filter out typing related issues before deployment.